test: dynamicTest 적용
